refactor: rename ledger popups
