+++ packages/shared/components/popups/LedgerNotConnectedPopup.svelte
